Add error handler for timeline and add convention for existing cases
This adds some error handling for invalid timelines being request and
adds a convention to allow selection of different timelines depending
on whether a case already exists or is new. If the case already exists,
the API will look for a timeline with '_existing' appended to the name.
If it exists, that timeline will be applied, otherwise the API sticks
to the requested timeline.